fix api extractor tags
