+++ packages/core/src/build-module.ts
@@ -1,4 +1,4 @@
-import { IgnitionError } from "./internal/errors";
+import { IgnitionError } from "./errors";
 import { ERRORS } from "./internal/errors-list";
 import { ModuleConstructor } from "./internal/module-builder";
 import { isValidIgnitionIdentifier } from "./internal/utils/identifier-validators";
