+++ packages/core/src/index.ts
@@ -1,14 +1,14 @@
 export { batches } from "./batches";
 export { buildModule } from "./build-module";
 export { deploy } from "./deploy";
-export * from "./internal/errors";
-export { ErrorDescriptor } from "./internal/errors-list";
+export * from "./errors";
 export { IgnitionModuleSerializer } from "./ignition-module-serializer";
 export { formatSolidityParameter } from "./internal/formatters";
 export { status } from "./status";
 export * from "./type-guards";
 export * from "./types/artifact";
 export * from "./types/deploy";
+export * from "./types/errors";
 export * from "./types/execution-events";
 export * from "./types/module";
 export * from "./types/module-builder";
