+++ packages/core/src/internal/deployer.ts
@@ -1,5 +1,6 @@
 import type { IgnitionModule, IgnitionModuleResult } from "../types/module";
 
+import { IgnitionError } from "../errors";
 import { isContractFuture } from "../type-guards";
 import { ArtifactResolver } from "../types/artifact";
 import {
@@ -18,7 +19,6 @@ import {
 
 import { Batcher } from "./batcher";
 import { DeploymentLoader } from "./deployment-loader/types";
-import { IgnitionError } from "./errors";
 import { ERRORS } from "./errors-list";
 import {
   initializeDeploymentState,
