+++ packages/core/src/internal/execution/abi.ts
@@ -7,9 +7,9 @@ import type {
   Result,
 } from "ethers";
 
+import { IgnitionError } from "../../errors";
 import { Artifact } from "../../types/artifact";
 import { ArgumentType, SolidityParameterType } from "../../types/module";
-import { IgnitionError } from "../errors";
 import { ERRORS } from "../errors-list";
 import { assertIgnitionInvariant } from "../utils/assertions";
