+++ packages/core/src/internal/execution/nonce-management/get-nonce-sync-messages.ts
@@ -1,5 +1,6 @@
 import uniq from "lodash/uniq";
 
+import { IgnitionError } from "../../../errors";
 import {
   isArtifactContractAtFuture,
   isNamedContractAtFuture,
@@ -10,7 +11,6 @@ import {
   IgnitionModule,
   IgnitionModuleResult,
 } from "../../../types/module";
-import { IgnitionError } from "../../errors";
 import { ERRORS } from "../../errors-list";
 import { getFuturesFromModule } from "../../utils/get-futures-from-module";
 import { getPendingOnchainInteraction } from "../../views/execution-state/get-pending-onchain-interaction";
