+++ packages/core/src/internal/validation/futures/validateArtifactLibraryDeployment.ts
@@ -1,8 +1,8 @@
+import { IgnitionError } from "../../../errors";
 import { isAccountRuntimeValue } from "../../../type-guards";
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { LibraryDeploymentFuture } from "../../../types/module";
-import { IgnitionError } from "../../errors";
 import { validateLibraryNames } from "../../execution/libraries";
 import { validateAccountRuntimeValue } from "../utils";
