+++ packages/core/src/internal/validation/futures/validateNamedStaticCall.ts
@@ -1,3 +1,4 @@
+import { IgnitionError } from "../../../errors";
 import {
   isAccountRuntimeValue,
   isArtifactType,
@@ -6,7 +7,6 @@ import {
 import { ArtifactResolver } from "../../../types/artifact";
 import { DeploymentParameters } from "../../../types/deploy";
 import { StaticCallFuture } from "../../../types/module";
-import { IgnitionError } from "../../errors";
 import { ERRORS } from "../../errors-list";
 import {
   validateArtifactFunction,
