+++ packages/core/src/status.ts
@@ -1,5 +1,5 @@
+import { IgnitionError } from "./errors";
 import { FileDeploymentLoader } from "./internal/deployment-loader/file-deployment-loader";
-import { IgnitionError } from "./internal/errors";
 import { ERRORS } from "./internal/errors-list";
 import { loadDeploymentState } from "./internal/execution/deployment-state-helpers";
 import { findDeployedContracts } from "./internal/views/find-deployed-contracts";
