+++ packages/core/src/verify.ts
@@ -1,6 +1,6 @@
+import { IgnitionError } from "./errors";
 import { builtinChains } from "./internal/chain-config";
 import { FileDeploymentLoader } from "./internal/deployment-loader/file-deployment-loader";
-import { IgnitionError } from "./internal/errors";
 import { ERRORS } from "./internal/errors-list";
 import { encodeDeploymentArguments } from "./internal/execution/abi";
 import { loadDeploymentState } from "./internal/execution/deployment-state-helpers";
